refactor: Add Copyright component
